Store Rc<Task> unconditionally in TaskOp::Move variant
The TaskOp::Move variant in its current form is a bit counter intuitive,
as it does not store the task in question but rather its index. That's
contrary to what we do for the TaskOp::Remove variant.
This change reworks TaskOp::Move to first and foremost store the task,
in an unconditional manner, and then optionally remember the position
that it was at before it got moved.