Introduce auxiliary data to the Db type
With upcoming changes we would like the ability to associate additional
data with items in the database. To make that happen we could just store
(T, Data) tuples in the Db to begin with, but that has trickle down
effects on consuming code, which now needs to be aware of this data,
just to ignore it (most likely). That would mean that we likely would
need to re-introduce a custom comparator to the Db as well, in order to
ignore this additional data when looking up items.
We want to try and avoid doing that and so with this change we embark on
a different route and introduce the concept of auxiliary data to the
database, building awareness of said additional data into the type
directly. The data is accessible via the Entry type and can be mutated
even on shared references by virtue of it being stored in a Cell.