Make db::Entry type aware of Db's Rc usage
When a Db is instantiated for a specific type T, we automatically add
the required Rc wrapper around it, as that is how the Db stores its
data. That's also how the db::Iter typedef behaves. The db::Enrty type,
however, is instantiated with Rc<T> directly, as it has no knowledge of
Rc usage itself. That makes it confusing to work with this type.
This change cleans up this inconsistency, adding direct awareness of Rc
usage to db::Entry. As a result, it can now instantiated in the same
with as Db and Iter.