Make db::Entry work with an index internally
So far, the db::Entry type stored a reference to the "slot" in the
database's vector of data. With upcoming changes we would like to
"chain" Entry objects, in order to navigate to the previous or next one,
in which case this approach is no longer flexible enough.
To fix this shortcoming, with this change we adjust the Entry type to
store a reference not just to the slot, but to the entire vector of the
database's data instead, finding the corresponding item in there with an
additional index member. We will later be able to adjust this index to
find the previous or next entry.