Introduce db::Entry::{next,prev} methods
Now that db::Entry is index based, this change introduces two additional
methods, next() and prev(), that allow instantiation of the next and
previous entry, respectively. We are now very close to having Entry be
an actual iterator, but just conceptually, the type is not designed to
provide iteration capabilities at its core. Rather, this new
functionality is mostly intended to be used for peeking one element back
or forth.