Re-use generated tasks in state tests
Some of the tests of the state module make the implicit assumption that
two calls of the make_tasks() helper with the same input arguments yield
the same task objects. Currently, that is a valid fact to assume, but
that is only due to a kludge in the ser::Task type's PartialEq
implementation, which excludes the Task's ID from equality checks.

We would like (and indeed need to) clean up this unclean design and as a
first step we convert the tests in the state module which make this
assumption to recreate the expected task state from the initially
created set of tasks, instead of basing it on newly created ones.