Keep task vectors around in tasks tests
Similar to the state module tests, some of the tests of the tasks module
assume that a serializable task object created with the same input
arguments is equal to an instance created earlier.
This change removes that assumption from the code, by creating a set of
tasks only once and then deriving the expected modified set based on it.