Work with task summaries in termui tests
Similar to tests of the state and tasks modules, a lot of the termui
tests implicitly assume that tasks created with the same input arguments
will be equal -- an assumption that is currently valid but that we would
like to challenge moving forward.
To stop relying on this behavior, this change switches all tests making
this assumption over to comparing task summaries instead.