Use auto derived PartialEq impl for ser::Task type
For a while now the ser::Task type has carried a custom impl of
PartialEq that specifically excludes the id member from being taken into
consideration for equality checks. While that was not too much of an
issue so far, it will become a problem in the future, when we anticipate
to add more members.
Luckily for us, by now all tests that implicitly assumed equality of two
tasks created with the same input arguments have been converted to work
with merely task summaries. That means that changing IDs are ignored by
them by design and, as such, we are free to include the id member in
equality checks for the type.
To that end, this change removes our bespoke PartialEq implementation
and replaces it with an auto-derive, which will do the right thing.