Introduce functionality for "distributed" tracking of relative positions
With recent changes we have implicitly enabled support for synchronizing
tasks with a server, potentially among many systems. This functionality
works fine, but it suffers from one major draw back: in order to
remember the overall order of tasks, currently we store a directory of
them in a special file containing meta data. Every time a task is added,
removed, or simply moved, this directory is updated. As a result,
synchronization between multiple systems can run into excessive amounts
of merge conflicts in this meta data file, unless there is a guarantee
that a local "checkout" works on the most up-to-date state and that
modifications can't happen concurrently.
Naturally, we do not want to make any such assumptions. So, in order to
remove this single point of "contention", we would like to annotate each
task with its position instead. The problem with doing that in a naive
way, we would still end up with potentially all tasks being updated if a
task is inserted as the very first one.
To fix this problem in a smarter way, we would like to store relative
positions instead of absolute ones. The idea is that when a task is
moved or inserted, as long as its position indicates between which two
tasks the new one lies, this is sufficient for the program to establish
the overall ordering of tasks.
To make this vision a reality, this change introduces the Position type,
which encapsulates the logic for calculating and working with relative
positions. It provides the means for finding a position lying between
two others, based on the idea that we can always find a real number
between two real numbers (unless they are equal, of course). Instead of
reals, we work with limited precision floating point values, but the
idea is the same. The limited precision aspect is not predicted to be a
problem in real world scenarios, but even if it did there are solutions
to them, so we do not worry too much about that right now.