Store Position as auxiliary data in task database
With this change we store objects of the previously introduced Position
type as the auxiliary data in our task database. We need those positions
to be available alongside tasks, but we do not want to store them inside
the Task object itself, for they conceptually do not belong there. As
such, the auxiliary data makes for a good location for keeping this
information. Note that right now we only store "dummy" values in here.
Subsequent changes will adjust the logic to calculate proper positions
as needed.