Simplify add_task() function
The add_task() function currently treats the cases where a target is
provided or not differently, using an insert into the database in one
case and a push to it in the other.
Future logic becomes more straightforward if we stick with inserting
unconditionally. Hence, this change refactors this function to do
precisely that.