Adjust Db::find() to return an Entry object
With db::Entry objects now storing an index of -- as opposed to a direct
reference to -- the referenced item in the database, we no longer need
to return only an index from the Db::find() method. Rather, we can
return a more versatile Entry object instead.
This change adjusts the method's signature accordingly and fixes up all
callers.