Maintain relative position data on a per-task basis
With this change we make sure to maintain correct relative task
positions in our Tasks type. In essence, whenever a task is added to the
database we update its relative position to lie between the positions of
its neighbor(s). Given that moving of tasks works by removing and adding
a task, adjusting this path takes care of all our position maintenance
needs.