Use per-task relative positions for re-establishing task ordering
So far we have used the task IDs stored in the TasksMeta type for
establishing the ordering of tasks after loading them from a directory.
By now we have all the pieces in place to rely on relative positions
stored on a per-task basis for this matter instead.
With this change we use said relative positions in order to establish
a well defined task ordering. If a task does not have a position set,
it will be ordered last and get an auto-assigned position based on the
previous task's position.