Add script for converting configuration from 0.2.* format to 0.3
Since version 0.2.6 the program's configuration format has changed
significantly. Staying compatible with the previous format long term was
never a goal and a manual conversion between the two formats is tedious
and best and borderline impossible at worst -- depending on the number
of tasks being managed.
The conversion can be performed by following the steps taken during
development. This change adds a script that formalizes these steps to
arrive at at least a poor-man's solution to semi-automating the process.