Remove built-in support for output redirection
A while back we added support for redirecting the program's output to a
file provided as a program argument. While that functionality is
working, there is an issue with it: we use termion::terminal_size() to
retrieve the terminal size, based on which we render the terminal UI.
This function assumes a global terminal and does not honor any inputs.
As such, its output is unlikely to match up with the actual terminal
dimensions when output redirection is active.
Given this issue, this change removes the feature. It turns out that the
original intent of easing debugging can be achieve by using gdb's --tty
argument directly:
  $ rust-gdb --tty /dev/pts/11 notnow