Add support for --version/-V option
This change adds support for a first optional argument to the program,
--version/-V. If set, the programs version (including a git revision, if
available) will be printed to stdout and the program will exit. In order
to determine the git revision at which the program was built, we rely on
the newly added grev build dependency.
Argument parsing is currently extremely bare-bones, because we only
support a single argument anyway. If we add more, we may want to
consider relying on a crate such as clap instead.