Abstract away stdin usage
While we pass in a "handle" to stdout to the run_prog() function, stdin
is not actually passed through. Rather, the receive_keys() function just
uses it directly.
In an attempt to abstract its usage away better, this change
parametrizes run_progs() with a reader that is passed to receive_keys()
eventually.