Add "end-to-end" test
Code coverage information has made it clear that we lack tests for
virtually all functionality provided by lib.rs.
To address this issue at least to some degree, this change introduces a
test invoking the run_prog() function, which captures a fair amount of
the contents of said file. As it turns out, doing so actually uncovered
a bug in our TermRenderer type, where we were potentially underflowing
an unsigned integer.