Only allow a single running program instance
While it is possible to run multiple instances of the program, doing so
is likely unintentional and almost certainly wrong: if one instance
saves modified state it will be overwritten by the other. In short,
given that we always work on a system-wide configuration (i.e., the user
cannot overwrite the configuration file or task directory), we cannot
think of a scenario where running more than one program instance is NOT
wrong.
To prevent that from happening, this change makes sure that we attempt
to "acquire" (create) a lock file before actually starting up. If the
file already exists, we error out. Otherwise we leave it hanging around
for the lifetime of the program and remove it once done. The behavior
can be overwritten via the newly introduced --force/-f arguments, which
can be useful if, for one reason or another, a stale lock file was left
lying around.