Make ToSerde trait use an associated type
The generic parameter of the ToSerde trait is really used to define the
output or target type. For such a use-case, however, it is not very well
suited: we can't even name it from within a definition but have to
repeat ourselves.
Associated types cater to precisely this use-case of an implementation
defining one and only one target type. As such, it makes sense for us to
to use an associated type as opposed to a generic parameter to specify
the target of the conversion. With this change we adjust the trait and
its clients accordingly.