Rework TaskState::with_serde() constructor
We try to have a 1:1 relation ship between "actual" program types and
those used for serialization and deserialization. Further, we strive for
having a XXX::with_serde() constructor for such program type objects
that takes as input an object of the serializable type.
In its current incarnation, the TaskState::with_serde() constructor
violates this unwritten contract: instead of accepting a
ser::state::TaskState it works with a ser::tasks::Tasks and
ser::tags::Templates combination.
This change fixes this matter, adjusting the constructor to expect a
ser::state::TaskState object instead.