Make TestUi::load_state() return a UiState object
We only ever use the TestUi::load_state() helper method to check
properties of the UI state, but not the task state. Yet, the State
object it returns comprises both kinds of states.
This change renames the method to load_ui_state() and adjusts it to only
return the UI state to callers.