Implement State in terms of TaskState and UiState
Our State type is, in many ways, just gluing together together TaskState
and UiState. However, currently it adds a bunch of logic for loading
objects of said types in unison.
On closer inspection it is pretty clear that this functionality actually
should reside in the TaskState and UiState types directly instead.
Hence, this change moves a bunch of this loading logic into newly
introduced TaskState::load() and UiState::load() constructors, together
with a UiState::with_serde() one mirroring what TaskState already had.
As a result of this refactoring, both State constructors are now trivial
in nature -- merely bundling calls to the TaskState and UiState
constructors.