Remove tasks_root member from TaskState type
The storing of the path to the task directory inside the TaskState type
has always felt a bit odd: the path itself is not really part of the
state that is captured by this type in a strict sense, it is more meta
data about it. That has led to shenanigans such as TaskState::with_serde()
requiring a path to be provided for no other reason than to store it so
that we can save the state back later on.
With this change we rework this arrangement: instead of storing the path
inside TaskState, we move it into TermUiData. This data is available
when saving the state back, and so it is very easy to provide the path
to that operation.