Remove path member from UiState type
Similar to the removal of the tasks_root member from the TaskState type,
this change removes the path member from UiState. Its purpose is similar
and so is the reasoning for removing it: it makes more sense for the
path to be provided by TermUiData when performing a save operation, as
opposed to storing it as part of the UiState object itself.