Introduce and use FilePath typedef
With upcoming changes we will need to access the directory part of a
path separately from the file name. While the standard Path and PathBuf
types support doing so via the parent() and file_name() methods, both
are fallible. Hence, it makes more sense for us to keep directory and
file name separate, and only combine them into the final path when that
is truly what is needed.
This change introduces the FilePath typedef, which represents such a
(directory, file name) tuple. We return that from the ui_config()
function and store it instead of a PathBuf.