Introduce capability infrastructure for guarding directory accesses
It is entirely possible for a user to modify tasks outside of the
program's control, but while it is running. So far this has not really
been too much of a cause for concern, because a subsequent write would
overwrite the changes made and the world would move on.
While that is still the case conceptually, with the recently added
support for synchronization of tasks with a server the stakes are higher
while the risk for such accidental mishaps has increased. The risk
increased, because users are now more likely to synchronize with a
server while the program is running. The consequences of such an action
are more severe, because if the program overwrites changes pulled in via
a task synchronization, the newly written state would be conveyed to the
server in a subsequent sync -- but it really represents an outdated view
of the world!
To curb such accidental misuse, this change introduces the
infrastructure for restricting the permissions of the tasks folder (and
more generally any folder) while the program is running, while granting
write access if a "token" is presented.
More precisely, we introduce the concept of a capability for writing to
the contents of a directory. While the capability exists, the directory
will be read-only and only accesses via the capability will allow
modifications.
Note, however, that this infrastructure is solely aimed at preventing
accidental misuse. It does nothing more than adjust file permissions and
users still have the power to revert those, given that they own the
files -- but by default a missing write permission will prevent
modifications.
This change only introduces the general infrastructure. It does not yet
hook it up with the program proper itself.