Hook up capability infrastructure
This change hooks up the previously introduced capability infrastructure
into the program. Really the most significant change is that to the
save_state_to_file() function, which now requires a FileCap object to be
presented to it. Because of that, and because this is effectively the
only function suitable for persisting state, all state save operations
now require such a capability to be provided. As a corollary, we are
sure that all state managed by the program is protected against
accidental modification outside the program's control, while the program
is running.
All other changes in this patch exist just to make the passing of a
FileCap to save_state_to_file() happen in the most sensible way.