Support editing of overly long tasks
When a task summary does not fit on the screen due to terminal size
limits, we display the first characters that fit and cut off the rest.
So far we have taken the same approach also when editing tasks. But of
course there the behavior ends up just being confusing: editing would
happen outside of what is being displayed and so the user is effectively
flying blind.
This change fixes this problem by supporting "scrolling" through overly
long task summaries when editing. The basic mechanism for doing that
within the renderer is the same as we employ for the TaskListBox's
selection: we maintain an offset that indicates the starting index of
the string to start rendering it at.