Add positive tag literals as tags to newly added task on empty view
When a view (or tab) currently has no tasks on it, adding a new task
cannot use the currently selected task as the "template" to copy tags
from. What currently happens is that the task is simply created without
any tags at all. That may result in a potentially horrible user
experience: if the user has no view configured that displays tasks
without any tags, the task will effectively end up in nirvana. The only
way to fix that is to manually edit the task's configuration file and
add the desired tags.
To fix this problem, this change makes sure to add all tags used as
positive tag literals in the view the task is added to (if the view is
empty). That may result in a few potentially unnecessary tags to be
present (depending on the complexity of the user's view configuration),
but it will result in the newly added task being displayed on the very
view in question, which allows for quick modification of the tags. While
this solution is not perfect (it is still only a heuristic: there is no
"right" way to to about it, aside from perhaps forcing the user to
select the tags to be present as part of task addition), it results in a
strictly better user experience compared to what we currently have.