Fix --version/-V option to work again
When we introduced the --force/-f option to the program, we missed the
fact that we were now effectively doing mutating operations on multiple
branches of a match statement on one and the same iterator. That broke
the version reporting via --version/-V, because the check for --force/-f
already consumed the iterator.
This change fixes the logic by making sure that we use a new iterator in
each branch of the match statement.