Adjust other task positions when failing to find "free" one
When we insert a new task into our database, we need to find a "free"
position that lies between the positions of the preceding and following
tasks. Given that a position is a floating point value with limited
precision, it is possible that no position can be found lying between
two others.
So far we have considered triggering this case highly unlikely, as it
involves repeatedly swapping tasks with each other or similar more or
less synthetic patterns. However, it is actually entirely possible to
trigger this case given enough time. E.g., if a user repeatedly adds a
new task on a tab that is configured to not show all tasks, what may end
up happening is that the new task is always added between the previously
added one and another one on a different tab. Eventually, there will not
be any space in between the two.
To fix issues like this once and for all, this change adjusts the
position finding logic to adjust the positions of follow on tasks if
there is no "space" between two. Doing so will allow us to always find a
valid position (barring the case where we exhaust the entire space of
representable f64 values).