Rename UI state to UI config
Currently we capture two concepts under a single name: UI state
comprises the UI's configuration (what the user desired) and its more or
less disposable state, such as the current tab or task selection. While
the state can basically be discarded at will and without much of an
impact on the user, the configuration very much should be preserved.

Currently, we effectively treat everything as configuration. In the
future we may be less stringent with disposable state. As a first step
towards distinguishing the two, this change renames UI state to UI
config.