Separate state test mixing task state and UI config
We have a bunch of tests in the state module that test both the working
of task state and that of UI config, in a single test case. Given that
we want to isolate the two concepts more from each other, this change
splits these tests into ones that test task state logic and others that
test UI configuration.