Privatize sub-modules of ui module
Currently we have pretty little isolation between components of the
program and while we do have a fairly layered design, it is easy for one
component or module to access functionality from another one, even if
that goes against said layering. The ui module is a perfect example of
that, as it exposes all its constituent modules publicly.
This change reworks exactly that, making all sub-modules in it private
and exporting only the types that are needed by clients. Doing so
enforces our layering much more cleanly and verifiably so.