Export types instead of modules for testing purposes
Our end-to-end test as well as the terminal UI example both peek into a
bunch of program internals for their reuse. So far this was made
possible by exposing the containing modules publicly. That is an overly
broad exposure and it leaks the internal structure of the program to the
"outside".
This change takes a different route, exposing only the types strictly
necessary to be exposed, but keeping their containing modules private.