Factor out ui::config module
Now that we have untangled task state from UI configuration, this change
factors out the UI configuration logic into a new sub-module inside the
ui module, config. There perhaps is an argument to be made that at least
part of the configuration should be "UI-agnostic", but going down the
route is unnecessary at this point, given that we only have a single UI
anyway.