Use "ui_config" nomenclature more in more places
We have a bunch of locations where we assume, by virtue of our variable
names, that only a single UI related configuration file/directory/entity
can exist.
With upcoming changes we plan to introduce the concept of "volatile" UI
state, in addition to the existing UI configuration. To make it easier
to differentiate between the two in the future, this change prefixes
variable names in locations where there there will arise ambiguity with
"ui_config_".