Split up UI configuration
Our UI configuration contains both user-configured bits as well as more
or less volatile state representing the current UI selections. Clubbing
these two concepts into a single object and using a single file to
persist them is sub-optimal: there is basically irrelevant information
exposed to users that they likely want to ignore and, more importantly,
it means that we pretty much end up saving the configuration every time
that the volatile state changed. If users opted to version control the
configuration, they will also witness a lot of noise caused by innocuous
selection changes.
With this change we split the two concepts, introducing UI state in
addition to our UI configuration. The state resides in the user's cache
directory (which is typically treated as discardable), whereas the
configuration stays in the config directory.