Store toggle_tag in TermUiData
When we introduced the "toggle tag" we added it directly as a member of
the TermUi type. That was done mostly for convenience, because the
friction for adding it to TermUiData, which really is the type that
should contain this kind of state, was greater.
This step has since been seen as flawed and with this change we correct
it.