Store colors in TermUi
When we load the program's configuration we also load the colors we
intend to use for the terminal based UI. However, given that nothing
currently changes these colors, we only pass them to the renderer and
then forget them. Among other things, this non-availability resulted in
some weird "warts" when saving our UI configuration back. Specifically,
what we ended up doing was reading the current configuration from disk
again, extract the colors, and only then save everything. That in turn
also forced us to wrap the colors in a Cell inside the Config type
itself.
With this change we make sure to store the colors inside the TermUi's
data itself. In so doing we not only mirror what we were already doing
with the "toggle tag" (which is stored there for pretty much the same
reason), it also affords us to destruct the Config object earlier,
and only pass truly needed configuration to the TermUi::new()
constructor.