Store "unwrapped" colors inside Config type
We currently store an Option of an instance of the Colors type inside
of Config objects. There no longer is a good reason for doing so: we
already load an object from the configuration and if none is present we
create a default instance.
Because we will always have an instance available, just remove the
Option wrapper altogether.