Warn if unsaved changes are present when trying to quit program
It has long been considered a problem that we simply allow quitting the
program while unsaved changes to tasks are present that would silently
be discarded in the process.
With this change we address this problem by issuing a "warning" and
intercepting the program exit. It is only after a second attempt of the
user to quit that we allow discarding of unsaved changes.