Introduce Line abstraction
The core of our input logic revolves around the InOut type's Input
variant, which marries the string to edit with a selection. However,
owing due to the standard library's String's interface, which provides
access to characters, but defaults to working with individual bytes
(e.g., in indexing operations or when retrieving its length), there is
potential for a lot of confusion as to whether a selection refers to a
byte or character index. This has resulted in bugs in the past (e.g.,
panics caused by String operations that implicitly panic on certain
inputs), but also is the cause of duplicated and confusing logic for
converting between the two "types" of indices.
To resolve problems like this, this change introduces the Line type. An
object of this type marries a string with the corresponding selection,
providing character based operations that are natural for clients to
reason about and that honor said selection while taking care of
correctly working with byte indexes internally. This change only
introduces the type. Subsequent changes will make use of it.