Correctly handle selection indexes for all Unicode inputs
Rust's String type works properly with Unicode inputs in the sense that
it allows iteration over characters. However, a character in its sense
does not necessarily meet the user's expectation of what a character is.
Strings that include variant selectors, for example, will be recognized
as multiple characters, when really they should be treated as one from a
user's point of view. For us that meant that the selection index was
wrong for such inputs.
With this change we add full Unicode awareness to the crate. We rely on
the newly introduced unicode-segmentation dependency to split inputs
based on grapheme clusters as opposed to "characters", which map to the
user's intuition.