Handle wide characters properly
As it turns out, there is even more to Unicode than we were aware of.
Specifically, Unicode has a notion of wide characters and those can span
multiple columns in a terminal. If such a character is displayed, we
once again end up with a wrongly displayed selection cursor.
This change fixes the problem by taking into account the width of
characters when calculating selection indexes.