Introduce --help/-h option for displaying help text
We strongly consider it back practice for a program to not support both
the long --help and short -h options. In the case of this application,
we currently support neither.
This change honors both options and displays a help text. The help text
layout is similar to what a crate like clap would produce, but we are
still reluctant to pull it in for our arguably simple argument handling
needs.