Add benchmarks for core Db operations
This change introduces a set of benchmarks for "core" operations on our
Db type. Specifically, we measure runtime performance of insertion in
various flavors, look up, and removal.
We start running these benchmarks in CI. While the environment does not
have particularly reliable performance characteristics, it can come in
handy to have ballpark numbers available frequently.
This work marks the start of some attempts to optimize performance of
this database.

> test db::tests::bench_finding      ... bench:  25,206,638 ns/iter (+/- 26,022)
> test db::tests::bench_insert       ... bench:   5,179,792 ns/iter (+/- 21,777)
> test db::tests::bench_push         ... bench:     345,502 ns/iter (+/- 434)
> test db::tests::bench_remove_first ... bench:   5,222,598 ns/iter (+/- 75,087)
> test db::tests::bench_remove_last  ... bench:     339,361 ns/iter (+/- 5,323)
> test db::tests::bench_try_insert   ... bench:  29,511,710 ns/iter (+/- 77,445)
> test db::tests::bench_try_push     ... bench:  25,208,024 ns/iter (+/- 27,784)