Introduce data base index for pointer based lookups
Our Db, while not particularly slow, is arguably not particularly
sophisticated when it comes to looking up elements by pointer.
Specifically, right now that is implemented as a linear scan over the
entire database (or at least until a match has been found).
In a first attempt to speed things up, this change introduces an index
to the database, which allows for faster lookups based on pointer
values.
The current implementation is based on a vector of Rc objects sorted by
pointer value. To map these back to indices, we store a "replayable"
index alongside, which we then adjust based on recorded insert and
remove operations on the database.

> test db::tests::bench_finding      ... bench:  25,481,506 ns/iter (+/- 155,643)
> test db::tests::bench_insert       ... bench:   6,828,397 ns/iter (+/- 28,376)
> test db::tests::bench_push         ... bench:   1,603,652 ns/iter (+/- 88,243)
> test db::tests::bench_remove_first ... bench:  17,725,903 ns/iter (+/- 263,343)
> test db::tests::bench_remove_last  ... bench:   1,026,021 ns/iter (+/- 6,556)
> test db::tests::bench_try_insert   ... bench:   5,482,184 ns/iter (+/- 28,053)
> test db::tests::bench_try_push     ... bench:     645,505 ns/iter (+/- 6,927)