Write back "replayed" indexes
Our index sped up some database operations, but it had little to no
effect on lookups -- which arguably were the primary target. The reason
for that behavior is that replaying indexes can be slow and do that
repeatedly.
That does not have to be the case. With this change we write back the
"replayed" index value to the index. Doing so improves performance for
repeated lookups potentially significantly.

> test db::tests::bench_finding      ... bench:     803,640 ns/iter (+/- 2,547)
> test db::tests::bench_insert       ... bench:   6,999,039 ns/iter (+/- 37,911)
> test db::tests::bench_push         ... bench:   1,608,979 ns/iter (+/- 3,575)
> test db::tests::bench_remove_first ... bench:  17,091,068 ns/iter (+/- 190,319)
> test db::tests::bench_remove_last  ... bench:   1,030,642 ns/iter (+/- 5,896)
> test db::tests::bench_try_insert   ... bench:   5,478,892 ns/iter (+/- 37,230)
> test db::tests::bench_try_push     ... bench:     633,891 ns/iter (+/- 8,451)