Use HashMap as index data structure
This change switches our database index over to using a HashMap a its
core data structure. The most present problem with the sorted vector was
that whenever we removed an element from the index, we had to shift all
subsequent ones. That is costly, but it could not really be avoided
given that any replacement with a sentinel would result in an unordered
sequence.
In order to work around this shortcoming, this change switches over to
using a HashMap as the core data structure for the index. In a HashMap,
removals can be implemented more efficiently and as a result we see
consistently low iteration times, without a spike for removals.

> test db::tests::bench_finding      ... bench:     517,878 ns/iter (+/- 40,414)
> test db::tests::bench_insert       ... bench:   6,821,373 ns/iter (+/- 100,534)
> test db::tests::bench_push         ... bench:   1,949,139 ns/iter (+/- 7,207)
> test db::tests::bench_remove_first ... bench:   6,499,591 ns/iter (+/- 34,965)
> test db::tests::bench_remove_last  ... bench:   1,559,242 ns/iter (+/- 2,197)
> test db::tests::bench_try_insert   ... bench:   6,906,861 ns/iter (+/- 29,445)
> test db::tests::bench_try_push     ... bench:   1,942,096 ns/iter (+/- 5,853)