Optimize representation of InsDel
The representation of the InsDel type is somewhat sub-optimal.
Specifically, because of the usage of an enum without the availability
of a niche inside the variant types, we end up with a total size of
eight bytes.
With this change we optimize said representation a bit, halving the size
of its instances.