Introduce task builder infrastructure
Creation of a task with specialized with_XXX constructors that accept a
certain number of arguments is not super expressive (given the unnamed
nature of function arguments) and not particularly scalable (in the
presence of further task attribute additions).
This change introduces a task builder infrastructure, that makes it
somewhat easier to create a Task object with certain attributes.