Use more obvious syntax for cloning tasks
We have several locations in which we need to make a deep copy of a Task
object. In general, our tasks are stored within an Rc and so we have to
clone the "deref" result.
In order to make it stand out more what is happening but also *ensure*
that it is indeed a Task object that is being cloned, switch over to
using the explicit Task::clone(..) syntax.