Rename Line to EditableText
So far the program only ever supported input of a single line of text.
That is going to change soon. Our 'Line' type actually does not
fundamentally prevent multi line text in it, it was just named
in line with the aforementioned observation.
To make it more obvious that said type cannot only handle single line
input, rename it to EditableText instead.