Factor out Text type
The EditableText type combines some Unicode aware text with a selection
and operations based on it. However, conceptually it makes sense to have
"non-editable" text, that just provides the necessary Unicode aware
building blocks without carrying an associated selection.
This change factors out this functionality as a separate type, Text.