Use more non-destructive methods for modifying an EditableText
The EditableText type has a slew of mutating methods that consume the
object and provide a new one. That was mostly done because it fit the
requirements of the caller best, which always expected a new object
anyway.
However, this is a very unconventional interface and it comes with a
bunch of drawbacks. As such, this change adjusts the type's API surface
to use mutating but non-destructive methods instead.