Introduce InputText type
This change introduces a type that encapsulates the existing input
handling that we use in the input/output widget, which is based on
terminal character sequences. Depending on whether readline support is
enabled or not, it either implements the evaluation of keys itself or
defers to libreadline.
Note that this change only introduces the type. It will be used
subsequently.