Use "cursor" terminology instead of "selection" in EditableText
Our usage of the word "selection" to refer to what effectively amounts
to a cursor is a bit confusing, as a selection in an input context
typically refers to something that is being replaced by input or
otherwise worked with and that can stretch multiple characters. In our
case, however, it is merely an index where to insert/remove the next
character.
With this change we rename "selection" to "cursor" in an attempt to
remove ambiguity.